Broke down transforms.affine into functionality-specific submodules
